backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
sqlparse==0.5.3
typing_extensions==4.13.1
gunicorn>=20.1
psycopg2-binary>=2.9
django-cors-headers>=4.7.0
djangorestframework>=3.16.0
Pillow>=11.1.0
django-ckeditor-5
django-jazzmin>=2.6.0
django-cleanup>=9.0.0
django-prometheus>=2.3.0
drf-yasg>=1.21.10
django-parler>=2.3
django-parler-rest>=2.2
django-solo>=2.4.0
requests
python-slugify
pre-commit
django-stubs==5.1.1
djangorestframework-stubs==3.15.2
types-requests